src/requirements.txt
boto3>=1.34.0
botocore>=1.34.0
Pillow>=10.0.0
numpy>=1.24.0